+++ ObservatoryFramework/Files/ParameterTypes/OutfittingModule.cs
@@ -2,7 +2,7 @@
 {
     public class OutfittingModule
     {
-        public int id { get; init; }
+        public ulong id { get; init; }
         public string Name { get; init; }
         public int BuyPrice { get; init; }
     }
